ssafy_book_project/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.7.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==4.2.5
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
exceptiongroup==1.2.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
numpy==2.2.5
openai==1.61.1
pillow==11.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.3.0
Wikipedia-API==0.8.1
